refactor

+++ app/src/main/java/com/pat/equalizer/model/BandLevel.kt
@@ -1,3 +1,3 @@
 package com.pat.equalizer.model
 
-data class BandLevel(val level: Float, val hzCenterFrequency: String)
+data class BandLevel(val level: Short, val hzCenterFrequency: String)
